adding test to fetch from the counter farm

+++ src/main/kotlin/com/dsw/autofill/AutoFillQueues.kt
@@ -2,7 +2,7 @@ package com.dsw.autofill
 
 interface AutoFillQueues<AutoFillQueue> {
 
-    fun addQueue(key: String,queue:AutoFillQueue,replenishThreshold:Float)
+    fun addQueue(key: String,queue:AutoFillQueue,replenishThreshold:Int)
     fun getObjects(key:String,count:Int):List<*>
     fun isQueueAvailable(key:String):Boolean
     fun getQueue(key:String): AutoFillQueue?
